requirements.txt
llama-index>=0.10.0,<1
llama-index-readers-file>=0.1.0,<1
llama-index-embeddings-fastembed>=0.5.0,<1
llama-index-retrievers-bm25>=0.1.0,<1
pillow>=10.3.0,<11
pypdf>=5.1.0,<7
python-docx>=1.1.0,<2
mcp>=1.0.0,<2
python-dotenv>=1.0.0,<2
huggingface-hub>=0.22.0,<2
flashrank>=0.1.0,<1
fastembed>=0.5.0,<1
pyyaml>=6.0,<7
